+++ packages/terra-table/tests/wdio/table-spec.js
@@ -1,38 +1,38 @@
 Terra.describeViewports('Table', ['medium'], () => {
-  describe('Fill', () => {
-    before(() => browser.url('/#/raw/tests/terra-table/table/fill'));
-    Terra.it.validatesElement();
+  it('displays table with Fill', () => {
+    browser.url('/raw/tests/terra-table/table/fill');
+    Terra.validates.element('fill');
   });
-  describe('HeaderCheckMark', () => {
-    before(() => browser.url('/#/raw/tests/terra-table/table/header-check-mark'));
-    Terra.it.validatesElement();
+  it('displays table with HeaderCheckMark', () => {
+    browser.url('/raw/tests/terra-table/table/header-check-mark');
+    Terra.validates.element('header check mark');
   });
-  describe('HeaderFooter', () => {
-    before(() => browser.url('/#/raw/tests/terra-table/table/header-footer'));
-    Terra.it.validatesElement();
+  it('displays table with HeaderFooter', () => {
+    browser.url('/raw/tests/terra-table/table/header-footer');
+    Terra.validates.element('header footer');
   });
-  describe('Padding', () => {
-    before(() => browser.url('/#/raw/tests/terra-table/table/padding'));
-    Terra.it.validatesElement();
+  it('displays table with Padding', () => {
+    browser.url('/raw/tests/terra-table/table/padding');
+    Terra.validates.element('padding');
   });
-  describe('Row', () => {
-    before(() => browser.url('/#/raw/tests/terra-table/table/row'));
-    Terra.it.validatesElement();
+  it('displays Row', () => {
+    browser.url('/raw/tests/terra-table/table/row');
+    Terra.validates.element('row');
   });
-  describe('Section', () => {
-    before(() => browser.url('/#/raw/tests/terra-table/table/section'));
-    Terra.it.validatesElement();
+  it('displays Section', () => {
+    browser.url('/raw/tests/terra-table/table/section');
+    Terra.validates.element('section');
   });
-  describe('Striped', () => {
-    before(() => browser.url('/#/raw/tests/terra-table/table/striped'));
-    Terra.it.validatesElement();
+  it('displays Striped table', () => {
+    browser.url('/raw/tests/terra-table/table/striped');
+    Terra.validates.element('striped');
   });
-  describe('Custom Cell Content', () => {
-    before(() => browser.url('/#/raw/tests/terra-table/table/custom-cell'));
-    Terra.it.validatesElement();
+  it('displays table with Custom Cell Content', () => {
+    browser.url('/raw/tests/terra-table/table/custom-cell');
+    Terra.validates.element('custom cell content');
   });
-  describe('Multi Demo', () => {
-    before(() => browser.url('/#/raw/tests/terra-table/table/multi-demo'));
-    Terra.it.validatesElement();
+  it('displays a Multi Demo table', () => {
+    browser.url('/raw/tests/terra-table/table/multi-demo');
+    Terra.validates.element('multi demo');
   });
 });